+++ hpc/nvidia/cuda/cuda_adapter.h
@@ -84,7 +84,7 @@
  * Pinned memory, on CPU memory, GPU accessable, DMA copy
  */
 #define cudaMallocHost(ptr, sz, flags)	__cuda(MallocHost(ptr, sz, flags))
-#define cudaFreeHost(ptr, sz, flags)	__cuda(FreeHost(ptr, sz, flags))
+#define cudaFreeHost(ptr)	__cuda(FreeHost(ptr))
 #define cudaMemset(ptr, v, size)	__cuda(Memset(ptr, v, size))
 /**
  * CUDA 12